pyproject.toml
[tool.poetry]
name = "aiowire"
version = "3.0.0"
description = "A simple event loop using asyncio"
authors = ["David M. Rogers <predictivestatmech@gmail.com>"]
readme = "README.rst"
repository = "https://github.com/frobnitzem/aiowire"
license = "LGPL-3.0-or-later"
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: AsyncIO"
]

[tool.poetry.dependencies]
python = ">=3.8"
pyzmq = "^22.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
pytest-asyncio = "^0.17.2"
pytest-cov = "^3.0.0"
mypy = "^1.5.1"

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = [ "tests" ]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
